+++ src/CommonUtils.h
@@ -7,8 +7,8 @@
 namespace CommonUtils{
 	int getGridXPoint(float point, nav_msgs::OccupancyGrid::ConstPtr map);
 	int getGridYPoint(float point, nav_msgs::OccupancyGrid::ConstPtr map);
-	int getTransformXPoint(int gridPoint, nav_msgs::OccupancyGrid::ConstPtr map);
-	int getTransformYPoint(int gridPoint, nav_msgs::OccupancyGrid::ConstPtr map);
+	float getTransformXPoint(int gridPoint, nav_msgs::OccupancyGrid::ConstPtr map);
+	float getTransformYPoint(int gridPoint, nav_msgs::OccupancyGrid::ConstPtr map);
 	int getIndex(int gridx, int gridy, nav_msgs::OccupancyGrid::ConstPtr map);
 }
 #endif
